Updates for docker.
